requirements.txt
Flask == 2.3.3
Flask-WTF == 1.1.1
WTForms == 3.1.2
email-validator == 2.1.0.post1
Werkzeug == 2.3.7
Flask-SQLAlchemy == 3.1.1
Flask-Login == 0.6.3
Flask-Migrate == 4.0.5
python-dotenv
pymysql
